actually

+++ controllers/cards.js
@@ -16,7 +16,7 @@ module.exports.getAllCards = (req, res) => {
     .catch((error) => {
       res
         .status(ERROR_CODE_SERVER_PROBLEM)
-        .json({ error: "Failed to fetch cards" });
+        .json({ message: "Failed to fetch cards" });
     });
 };
 
@@ -24,20 +24,6 @@ module.exports.getAllCards = (req, res) => {
 module.exports.createCard = (req, res) => {
   const { name, link } = req.body;
   const { _id } = req.user;
-
-  // if (!name || name.length < 2 || name.length > 30) {
-  //   return res.status(ERROR_CODE).json({
-  //     error: "Name should be between 2 and 30 characters long",
-  //     message: "name is less than 2 symbols or more than 30",
-  //   });
-  // }
-
-  // if (!link) {
-  //   return res.status(ERROR_CODE).json({
-  //     error: "You need to write link",
-  //     message: "You need to write link",
-  //   });
-  // }
 
   Card.create({ name, link, owner: _id })
     .then((card) => res.status(201).send(card))
@@ -49,7 +35,7 @@ module.exports.createCard = (req, res) => {
       } else {
         res
           .status(ERROR_CODE_SERVER_PROBLEM)
-          .json({ error: "Failed to create card" });
+          .json({ message: "Failed to create card" });
       }
     });
 };
@@ -58,25 +44,20 @@ module.exports.createCard = (req, res) => {
 module.exports.deleteCard = (req, res) => {
   const { cardId } = req.params;
 
-  // if (!mongoose.Types.ObjectId.isValid(cardId)) {
-  //   return res
-  //     .status(ERROR_CODE)
-  //     .json({ error: "Invalid card ID", message: "Wrong card id" });
-  // }
+  if (!mongoose.Types.ObjectId.isValid(cardId)) {
+    return res.status(ERROR_CODE).json({ message: "Wrong card id" });
+  }
 
   Card.findByIdAndDelete(cardId)
     .then((deletedCard) => {
       if (deletedCard) {
         res.json(deletedCard);
       } else {
-        res
-          .status(ERROR_CODE_NOT_FOUND)
-          .json({ error: "Card not found", message: "Wrong card id" });
+        res.status(ERROR_CODE_NOT_FOUND).json({ message: "Wrong card id" });
       }
     })
     .catch((error) => {
-      res.status(ERROR_CODE).json({
-        error: "Failed to delete card",
+      res.status(ERROR_CODE_SERVER_PROBLEM).json({
         message: "Deleting a card with an incorrect id",
       });
     });
@@ -103,7 +84,7 @@ module.exports.likeCard = async (req, res) => {
     if (!card) {
       return res
         .status(ERROR_CODE_NOT_FOUND)
-        .json({ error: "Card not found", message: "Wrong like id" });
+        .json({ message: "Wrong like id" });
     }
 
     // if (card.likes.includes(userId)) {
@@ -112,9 +93,6 @@ module.exports.likeCard = async (req, res) => {
     //     message: "Incorrect like id",
     //   });
     // }
-
-    card.likes.push(userId);
-    await card.save();
 
     res.status(200).json(card);
   } catch (error) {
@@ -125,7 +103,7 @@ module.exports.likeCard = async (req, res) => {
     } else {
       res
         .status(ERROR_CODE_SERVER_PROBLEM)
-        .json({ error: "Failed to like card" });
+        .json({ message: "Failed to like card" });
     }
   }
 };
@@ -152,17 +130,10 @@ module.exports.dislikeCard = async (req, res) => {
     if (!card) {
       return res
         .status(ERROR_CODE_NOT_FOUND)
-        .json({ error: "Card not found", message: "Wrong like id" });
+        .json({ message: "Wrong like id" });
     }
 
     const index = card.likes.indexOf(userId);
-
-    if (index === -1) {
-      return res.status(200).json({
-        error: "User has not liked this card.",
-        message: "Incorrect like id",
-      });
-    }
 
     card.likes.splice(index, 1);
     await card.save();
@@ -176,7 +147,7 @@ module.exports.dislikeCard = async (req, res) => {
     } else {
       res
         .status(ERROR_CODE_SERVER_PROBLEM)
-        .json({ error: "Failed to dislike card" });
+        .json({ message: "Failed to dislike card" });
     }
   }
 };
